Revert "async commit + wh (#263)"
This reverts commit 272d9c6e921ec50f757469dc510cdcf549d9ec70.
